Added continue mode to test script.

+++ run_integration_tests.sh
@@ -1,5 +1,31 @@
 #!/bin/sh
 
+args=$(getopt c $*)
+if [ $? -ne 0 ]; then
+    echo "Usage: run_integration_tests.sh [-c]"
+    echo ""
+    echo "where:"
+    echo "\t-c      Continue; don't delete build and build from scratch."
+    echo ""
+    echo ""
+    exit 1
+fi
+
+#by default, don't continue
+cont=0
+
+set -- $args
+while [ $# -ne 0 ]
+do
+    case "$1"
+    in
+        -c)
+            cont=1; shift;;
+        --)
+            shift; break;;
+    esac
+done
+
 if [ ! $(id -u) -eq 0 ]; then
     echo "This script must be run as root."
     exit 1
@@ -18,17 +44,28 @@ else
     echo "Set unprivileged user to $unprivileged_user."
 fi
 
-if [ -d build ]; then
-    echo "Cleaning up build directory."
-    rm -rf build
+if [ $cont -ne 1 ]; then
+    if [ -d build ]; then
+        echo "Cleaning up build directory."
+        rm -rf build
+    fi
 fi
-
-set -e
 
 echo "PKG_CONFIG_PATH = $PKG_CONFIG_PATH"
 
-sudo -E -u $unprivileged_user mkdir build
+set -e
+
+if [ $cont -ne 1 ]; then
+    sudo -E -u $unprivileged_user mkdir build
+    sudo -E -u $unprivileged_user meson subprojects update --reset
+fi
+
 cd build
-sudo -E -u $unprivileged_user meson ..
+if [ $cont -ne 1 ]; then
+    sudo -E -u $unprivileged_user meson ..
+else
+    sudo -E -u $unprivileged_user meson --reconfigure ..
+fi
+
 sudo -E -u $unprivileged_user ninja
 sudo -E -u $unprivileged_user ninja test
